requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.6
backports.entry-points-selectable==1.1.0
black==21.7b0
cfgv==3.3.0
click==8.0.1
distlib==0.3.2
Django==3.2.6
django-cors-headers==3.8.0
django-countries==7.2.1
django-model-utils==4.1.1
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.2
identify==2.2.13
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.2.0
pre-commit==2.14.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.6
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.21
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
virtualenv==20.7.2
wrapt==1.12.1
